Moving cli commands to seperate file
